# increase light throughput by 10%