--- a/mpf/core/light_controller.py
+++ b/mpf/core/light_controller.py
@@ -14,6 +14,9 @@ class LightController(MpfController):
 
     """Handles light updates and light monitoring."""
 
+    __slots__ = ["light_color_correction_profiles", "_initialised", "_monitor_update_task", "brightness_factor",
+                 "_brightness_template"]
+
     config_name = "light_controller"
 
     def __init__(self, machine: MachineController) -> None:
@@ -25,11 +28,19 @@ class LightController(MpfController):
 
         # will only get initialised if there are lights
         self._initialised = False
+        self._brightness_template = self.machine.placeholder_manager.build_float_template("machine.brightness", 1.0)
+        self._update_brightness()
 
         self._monitor_update_task = None                    # type: asyncio.Task
 
         if 'named_colors' in self.machine.config:
             self._load_named_colors()
+
+    def _update_brightness(self, *args):
+        """Update brightness factor."""
+        del args
+        self.brightness_factor, future = self._brightness_template.evaluate_and_subscribe([])
+        future.add_done_callback(self._update_brightness)
 
     def _load_named_colors(self):
         """Load named colors from config."""