Old (Commit c7a6bbec2d09bf2b05c3e658055b08f7219ab927):
Python 3.6.7: Baseline: 0.06651ms One LED: +0.35909ms 2784.81 fps; One LED (token): +0.36026ms 2775.77 fps; 30 LEDs: +1.97643ms 505.96 fps; Multi Step: +1.69915 588.53 fps;
PYPY 7.0: Baseline: 0.00610ms One LED: +0.03445ms 29029.25 fps; One LED (token): +0.03559ms 28097.30 fps; 30 LEDs: +0.17900ms 5586.74 fps; Multi Step: +0.18539 5393.95 fps;

New (Commit 223376d170428355bbad209998091f94b905a362):
Python 3.6.7: Baseline: 0.06479ms One LED: +0.33401ms 2993.89 fps; One LED (token): +0.33319ms 3001.29 fps; 30 LEDs: +1.69342ms 590.52 fps; Multi Step: +1.46176 684.11 fps;
PYPY 7.0: Baseline: 0.00602ms One LED: +0.03414ms 29293.65 fps; One LED (token): +0.03556ms 28121.22 fps; 30 LEDs: +0.18483ms 5410.51 fps; Multi Step: +0.19293 5183.36 fps;